third_party/WebKit/LayoutTests/bluetooth/characteristic/characteristicProperties.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return getHealthThermometerService()
    .then(({service}) => Promise.all([
      service.getCharacteristic('temperature_measurement'),
      service.getCharacteristic('measurement_interval')]))
    .then(([temperature_measurement, measurement_interval]) => {
      let tm_expected_properties = new TestCharacteristicProperties([
        'indicate']);
      assert_properties_equal(temperature_measurement.properties,
                              tm_expected_properties);

      let mi_expected_properties = new TestCharacteristicProperties([
        'read', 'write', 'indicate']);
      assert_properties_equal(measurement_interval.properties,
                              mi_expected_properties);
    });
}, 'HeartRate device properties');
</script>
